Refactor expression function management.
