+++ packages/vega-expression/index.js
@@ -1,4 +1,19 @@
-export { default as ASTNode } from './src/ast';
+export {
+  RawCode,
+  Literal,
+  Property,
+  Identifier,
+  ArrayExpression,
+  BinaryExpression,
+  CallExpression,
+  ConditionalExpression,
+  LogicalExpression,
+  MemberExpression,
+  ObjectExpression,
+  UnaryExpression,
+  default as ASTNode
+} from './src/ast';
+
 export { default as parse} from './src/parser';
 export { default as codegen } from './src/codegen';
 export { default as functions } from './src/functions';
