+++ packages/vega-functions/index.js
@@ -1,0 +1,108 @@
+export {
+  codegenParams,
+  codeGenerator,
+  expressionFunction,
+  functionContext
+} from './src/codegen';
+
+export {
+  data,
+  indata,
+  setdata
+} from './src/data';
+
+export {
+  default as encode
+} from './src/encode';
+
+export {
+  format,
+  utcFormat,
+  timeFormat,
+  utcParse,
+  timeParse,
+  monthFormat,
+  monthAbbrevFormat,
+  dayFormat,
+  dayAbbrevFormat
+} from './src/format';
+
+export {
+  geoArea,
+  geoBounds,
+  geoCentroid
+} from './src/geo';
+
+export {
+  default as inScope
+} from './src/inscope';
+
+export {
+  warn,
+  info,
+  debug
+} from './src/log';
+
+export {
+  default as merge
+} from './src/merge';
+
+export {
+  default as modify
+} from './src/modify';
+
+export {
+  pinchDistance,
+  pinchAngle
+} from './src/pinch';
+
+export {
+  DataPrefix,
+  IndexPrefix,
+  ScalePrefix,
+  SignalPrefix
+} from './src/prefix.js';
+
+export {
+  range,
+  domain,
+  bandwidth,
+  bandspace,
+  copy,
+  scale,
+  invert
+} from './src/scale';
+
+export {
+  default as scaleGradient
+} from './src/scale-gradient';
+
+export {
+  geoShape,
+  pathShape
+} from './src/shape';
+
+export {
+  treePath,
+  treeAncestors
+} from './src/tree';
+
+export {
+  containerSize,
+  screen,
+  windowSize
+} from './src/window';
+
+export {
+  dataVisitor,
+  indataVisitor,
+  scaleVisitor
+} from './src/visitors';
+
+export {
+  formatDefaultLocale as formatLocale
+} from 'd3-format';
+
+export {
+  timeFormatDefaultLocale as timeFormatLocale
+} from 'd3-time-format';
