+++ packages/vega-functions/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "vega-functions",
+  "version": "1.0.0",
+  "description": "Custom functions for the Vega expression language.",
+  "keywords": [
+    "vega",
+    "expression",
+    "functions"
+  ],
+  "author": "Jeffrey Heer (http://idl.cs.washington.edu)",
+  "license": "BSD-3-Clause",
+  "main": "build/vega-functions.js",
+  "module": "index.js",
+  "repository": "vega/vega",
+  "scripts": {
+    "rollup": "rollup -g d3-array:d3,d3-color:d3,d3-format:d3,d3-geo:d3,d3-time-format:d3,vega-dataflow:vega,vega-expression:vega,vega-scale:vega,vega-scenegraph:vega,vega-selections:vega,vega-statistics:vega,vega-util:vega -f umd -n vega -o build/vega-functions.js -- index.js",
+    "prebuild": "rimraf build && mkdir build",
+    "build": "yarn rollup",
+    "postbuild": "terser build/vega-functions.js -c -m -o build/vega-functions.min.js",
+    "pretest": "yarn prebuild && yarn rollup",
+    "test": "tape 'test/**/*-test.js' && eslint index.js src test",
+    "prepublishOnly": "yarn test && yarn build",
+    "postpublish": "git push && git push --tags"
+  },
+  "dependencies": {
+    "d3-array": "^2.0.3",
+    "d3-color": "^1.2.3",
+    "d3-format": "^1.3.2",
+    "d3-geo": "^1.11.3",
+    "d3-time-format": "^2.1.3",
+    "vega-dataflow": "^4.1.0",
+    "vega-expression": "^2.4.0",
+    "vega-scale": "^2.5.1",
+    "vega-scenegraph": "^3.2.3",
+    "vega-selections": "^1.0.0",
+    "vega-statistics": "^1.2.3",
+    "vega-util": "^1.7.1"
+  }
+}
