+++ packages/vega-functions/src/log.js
