+++ packages/vega-functions/src/merge.js
