+++ packages/vega-functions/src/pinch.js
@@ -1,12 +1,12 @@
 export function pinchDistance(event) {
-  var t = event.touches,
-      dx = t[0].clientX - t[1].clientX,
-      dy = t[0].clientY - t[1].clientY;
+  const t = event.touches,
+        dx = t[0].clientX - t[1].clientX,
+        dy = t[0].clientY - t[1].clientY;
   return Math.sqrt(dx * dx + dy * dy);
 }
 
 export function pinchAngle(event) {
-  var t = event.touches;
+  const t = event.touches;
   return Math.atan2(
     t[0].clientY - t[1].clientY,
     t[0].clientX - t[1].clientX
