+++ packages/vega-functions/src/prefix.js
@@ -1,0 +1,4 @@
+export const DataPrefix   = ':';
+export const IndexPrefix  = '@';
+export const ScalePrefix  = '%';
+export const SignalPrefix = '$';
