+++ packages/vega-parser/index.js
@@ -6,12 +6,6 @@ export {default as signalUpdates} from './src/parsers/signal-updates';
 export {default as stream} from './src/parsers/stream';
 
 export {
-  codeGenerator,
-  functionContext,
-  expressionFunction
-} from './src/parsers/expression/codegen';
-
-export {
   MarkRole,
   FrameRole,
   ScopeRole,
@@ -30,6 +24,3 @@ export {
 
 export {default as Scope} from './src/Scope';
 export {default as DataScope} from './src/DataScope';
-
-export {formatDefaultLocale as formatLocale} from 'd3-format';
-export {timeFormatDefaultLocale as timeFormatLocale} from 'd3-time-format';
