--- packages/vega-parser/src/parsers/expression/ast.js
@@ -1,2 +1,0 @@
-export var Literal = 'Literal';
-export var Identifier = 'Identifier';
