--- packages/vega-parser/src/parsers/expression/prefixes.js
@@ -1,3 +1,0 @@
-export var indexPrefix  = '@';
-export var scalePrefix  = '%';
-export var dataPrefix   = ':';
