+++ packages/vega-selections/README.md
@@ -1,0 +1,3 @@
+# vega-selections
+
+[Vega expression language](https://vega.github.io/vega/docs/expressions/) functions for implementing Vega-Lite selections.
