+++ packages/vega-selections/index.js
@@ -1,0 +1,3 @@
+export {selectionTest} from './src/selectionTest';
+export {selectionResolve} from './src/selectionResolve';
+export {selectionVisitor} from './src/selectionVisitor';
