+++ packages/vega-selections/src/constants.js
@@ -1,0 +1,2 @@
+export const Intersect = 'intersect';
+export const Union = 'union';
