+++ packages/vega-util/index.js
@@ -33,16 +33,25 @@ export {
   zoomSymlog
 } from './src/transform';
 
+export {
+  quarter,
+  utcquarter
+} from './src/quarter';
+
 export {default as array} from './src/array';
+export {default as clampRange} from './src/clampRange';
 export {default as compare} from './src/compare';
 export {default as constant} from './src/constant';
 export {default as debounce} from './src/debounce';
 export {default as error} from './src/error';
 export {default as extend} from './src/extend';
+export {default as extent} from './src/extent';
 export {default as extentIndex} from './src/extentIndex';
 export {default as fastmap} from './src/fastmap';
 export {default as field} from './src/field';
+export {default as flush} from './src/flush';
 export {default as inherits} from './src/inherits';
+export {default as inrange} from './src/inrange';
 export {default as isArray} from './src/isArray';
 export {default as isBoolean} from './src/isBoolean';
 export {default as isDate} from './src/isDate';
@@ -56,6 +65,7 @@ export {default as merge} from './src/merge';
 export {default as pad} from './src/pad';
 export {default as peek} from './src/peek';
 export {default as repeat} from './src/repeat';
+export {default as span} from './src/span';
 export {default as splitAccessPath} from './src/splitAccessPath';
 export {default as stringValue} from './src/stringValue';
 export {default as toBoolean} from './src/toBoolean';
