+++ packages/vega-util/src/span.js
@@ -1,0 +1,9 @@
+import peek from './peek';
+
+/**
+ * Return the numerical span of an array: the difference between
+ * the last and first values.
+ */
+export default function(array) {
+  return (peek(array) - array[0]) || 0;
+}
