+++ packages/vega-view/src/runtime.js
@@ -1,5 +1,5 @@
 import {transforms} from 'vega-dataflow';
-import {functionContext} from 'vega-parser';
+import {functionContext} from 'vega-functions';
 import {parse, context} from 'vega-runtime';
 
 export default function(view, spec, functions) {
