+++ packages/vega/index.d.ts
@@ -1,7 +1,3 @@
 declare module 'vega' {
   export * from 'vega-typings';
 }
-
-declare module 'vega-lib' {
-  export * from 'vega-typings';
-}
