+++ packages/vega/package.json
@@ -34,6 +34,7 @@
     "vega-encode": "^3.2.2",
     "vega-event-selector": "^2.0.0",
     "vega-expression": "^2.4.0",
+    "vega-functions": "^1.0.0",
     "vega-force": "^3.0.0",
     "vega-geo": "^3.1.1",
     "vega-hierarchy": "^3.1.0",
